move dml to ast conversion into own module
