+++ libs/datamodel/core/src/validator/directive/default.rs
@@ -1,6 +1,5 @@
 use crate::error::DatamodelError;
 use crate::validator::directive::{Args, DirectiveValidator};
-use crate::validator::LowerDmlToAst;
 use crate::{ast, dml, ValueGenerator};
 use prisma_value::PrismaValue;
 
@@ -56,13 +55,39 @@ impl DirectiveValidator<dml::Field> for DefaultDirectiveValidator {
         if let Some(default_value) = field.default_value() {
             return Ok(vec![ast::Directive::new(
                 self.directive_name(),
-                vec![ast::Argument::new(
-                    "",
-                    LowerDmlToAst::lower_default_value(default_value.clone()),
-                )],
+                vec![ast::Argument::new("", lower_default_value(default_value.clone()))],
             )]);
         }
 
         Ok(vec![])
+    }
+}
+
+pub fn lower_default_value(dv: dml::DefaultValue) -> ast::Expression {
+    match dv {
+        dml::DefaultValue::Single(v) => lower_prisma_value(&v),
+        dml::DefaultValue::Expression(e) => {
+            let exprs = e.args.iter().map(lower_prisma_value).collect();
+            ast::Expression::Function(e.name, exprs, ast::Span::empty())
+        }
+    }
+}
+
+pub fn lower_prisma_value(pv: &PrismaValue) -> ast::Expression {
+    match pv {
+        PrismaValue::Boolean(true) => ast::Expression::BooleanValue(String::from("true"), ast::Span::empty()),
+        PrismaValue::Boolean(false) => ast::Expression::BooleanValue(String::from("false"), ast::Span::empty()),
+        PrismaValue::String(value) => ast::Expression::StringValue(value.clone(), ast::Span::empty()),
+        PrismaValue::Enum(value) => ast::Expression::ConstantValue(value.clone(), ast::Span::empty()),
+        PrismaValue::DateTime(value) => ast::Expression::StringValue(value.to_rfc3339(), ast::Span::empty()),
+        PrismaValue::Float(value) => ast::Expression::NumericValue(value.to_string(), ast::Span::empty()),
+        PrismaValue::Int(value) => ast::Expression::NumericValue(value.to_string(), ast::Span::empty()),
+        PrismaValue::Null(_) => ast::Expression::ConstantValue("null".to_string(), ast::Span::empty()),
+        PrismaValue::Uuid(val) => ast::Expression::StringValue(val.to_string(), ast::Span::empty()),
+        PrismaValue::Json(val) => ast::Expression::StringValue(val.to_string(), ast::Span::empty()),
+        PrismaValue::List(vec) => ast::Expression::Array(
+            vec.iter().map(|pv| lower_prisma_value(pv)).collect(),
+            ast::Span::empty(),
+        ),
     }
 }
