+++ libs/datamodel/core/src/validator/dml_to_ast/mod.rs
@@ -1,0 +1,3 @@
+mod lower;
+
+pub use lower::LowerDmlToAst;
