+++ libs/datamodel/core/src/validator/mod.rs
@@ -1,17 +1,15 @@
-mod dml_to_ast;
+pub mod ast_to_dml;
+mod common;
+mod directive;
+pub mod dml_to_ast;
 mod invalid_model_names;
 mod lift;
-mod lower;
 mod precheck;
 mod standardise;
 mod validate;
 mod validation_pipeline;
 
-mod common;
-pub mod directive;
-
 use lift::*;
-pub use lower::*;
 use standardise::*;
 use validate::*;
 pub use validation_pipeline::*;
